rockpro64: dts: rename backlight_en to lcd_en
For consistency with the schematic.

As this is routed to pin 22 of the 30-pin eDP connector (which is
defined as BL_ENABLE in table 5-3 of the eDP 1.2 specification
if you believe page 9 of https://www.vesa.org/wp-content/uploads/2010/12/DisplayPort-DevCon-Presentation-eDP-Dec-2010-v3.pdf)
I'm leaving this attached to the backlight's enable-gpios.